+++ frontend/src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import userReducer from "../features/user/userSlice";
-import notificationsReducer from "../features/notifications/notificationsSlice";
+import userReducer from "../features/userSlice";
+import notificationsReducer from "../features/notificationsSlice";
 
 export default configureStore({
   reducer: {
