+++ frontend/src/components/pages/Login.js
@@ -6,7 +6,7 @@ import { useDispatch } from "react-redux";
 import { EyeInvisibleOutlined, EyeTwoTone } from "@ant-design/icons";
 import { LOGIN_ERROR_MSGS, ACCOUNT_TYPE } from "utils/consts";
 import { login, sendVerificationEmail } from "utils/auth.service";
-import { fetchUser } from "features/user/userSlice";
+import { fetchUser } from "features/userSlice";
 import MenteeButton from "../MenteeButton";
 import usePersistedState from "utils/hooks/usePersistedState";
 import "../css/Login.scss";
