+++ frontend/src/components/pages/Videos.js
@@ -6,7 +6,7 @@ import MentorVideo from "../MentorVideo";
 import VideoSubmit from "../VideoSubmit";
 import { SPECIALIZATIONS } from "utils/consts.js";
 import { formatDropdownItems } from "utils/inputs";
-import { updateAndFetchUser } from "features/user/userSlice";
+import { updateAndFetchUser } from "features/userSlice";
 import { ACCOUNT_TYPE } from "utils/consts.js";
 import useAuth from "utils/hooks/useAuth";
 import "../css/Videos.scss";
